+++ cli/src/connect/__test__/e2e/e2e_migrate_command/react_storybook/StorybookComponent.stories.tsx
@@ -1,0 +1,19 @@
+import { StoryParameters } from '../../../../../react/index_react'
+import { StorybookComponent } from './StorybookComponent'
+import React from 'react'
+
+export default {
+  title: 'StorybookComponent',
+  component: StorybookComponent,
+  parameters: {
+    design: {
+      type: 'figma',
+      url: 'https://figma.com/test',
+      examples: ['Default'],
+    },
+  } satisfies StoryParameters,
+}
+
+export function Default() {
+  return <StorybookComponent disabled={false}>Hello</StorybookComponent>
+}