+++ cli/src/connect/__test__/e2e/e2e_migrate_command/react_storybook/StorybookComponent.tsx
@@ -1,0 +1,11 @@
+import { ReactNode } from 'react'
+import React from 'react'
+
+export interface Props {
+  disabled: boolean
+  children: ReactNode
+}
+
+export const StorybookComponent = ({ disabled, children }: Props) => {
+  return <button disabled={disabled}>{children}</button>
+}